Refactor: refactor slider component and place detail page 슬라이더와 장소 세부정보 리팩토링
