+++ alcoholmap/src/frontend/src/components/commons/placedetail/PlaceDetail.container.tsx
