+++ alcoholmap/src/frontend/src/components/commons/placedetail/PlaceDetail.types.ts
