+++ alcoholmap/src/frontend/src/components/commons/ratedplace/RatedPlaceSlider.presenter.tsx
