+++ alcoholmap/src/frontend/src/components/commons/ratedplace/RatedPlaceSlider.styles.tsx
