+++ alcoholmap/src/frontend/src/components/commons/ratedplace/RatedPlaceSlider.types.ts
