+++ alcoholmap/src/frontend/src/components/commons/wishlist/WishListSlider.container.tsx
