+++ alcoholmap/src/frontend/src/components/commons/wishlist/WishListSlider.presenter.tsx
