+++ alcoholmap/src/frontend/src/components/commons/wishlist/WishListSlider.types.ts
